+++ config/production.py
@@ -2,3 +2,6 @@ from config.base import *
 
 # Flask
 DEBUG = False
+
+# Taskflow
+TASKFLOW_LOG_LEVEL = 'INFO'